requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.7.0
botocore==1.31.82
branca==0.7.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.6.4
distro==1.8.0
dnspython==2.4.2
ecdsa==0.18.0
fastapi==0.103.1
ffmpy==0.3.1
filelock==3.13.1
folium==0.15.1
fonttools==4.44.0
frozendict==2.3.10
frozenlist==1.4.1
fsspec==2023.10.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.1
httpx==0.25.1
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.0
iniconfig==2.0.0
install==1.3.5
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.1.5
langchain-community==0.0.19
langchain-core==0.1.21
langchain-openai==0.0.5
langsmith==0.0.87
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.1
mdurl==0.1.2
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
numpy==1.26.1
oauthlib==3.2.2
openai==1.11.1
orjson==3.9.10
packaging==23.2
pandas==2.1.2
passlib==1.7.4
patsy==0.5.3
peewee==3.17.0
Pillow==10.1.0
playwright==1.41.2
pluggy==1.3.0
protobuf==4.21.12
pyarrow==15.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydeck==0.8.1b0
pydub==0.25.1
pyee==11.0.1
pyffx==0.3.0
Pygments==2.16.1
PyJWT==1.7.1
pymongo==4.6.0
PyMySQL==1.1.0
pyparsing==3.1.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
s3transfer==0.7.0
scipy==1.11.4
seaborn==0.13.0
semantic-version==2.10.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
statsmodels==0.14.0
streamlit==1.31.0
streamlit-folium==0.18.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.23.2
validators==0.22.0
webencodings==0.5.1
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0
